Initial commit
This is designed to work with a computer hooked to the USB port (Serial) and another serial device on pins 2 and 3 (BTSerial).  

The original testing device was a BlueTooth module.  I was getting garbled text over 120-130 characters.  I think this was due to a memory issue with an ATMega 328, so instead of BTSerial.readString(), I'm using a read() method to bring in the string.

I'm assuming there is not any nested JSON objects, so the home grown method kills it's loop after any "}" character.